fix compilation errors
