+++ speeduino/cancomms.h
@@ -12,28 +12,28 @@
 #define NEW_CAN_PACKET_SIZE   75
 #define CAN_PACKET_SIZE   75
 
-uint8_t currentcanCommand;
-uint8_t currentCanPage = 1;//Not the same as the speeduino config page numbers
-uint8_t nCanretry = 0;      //no of retrys
-uint8_t cancmdfail = 0;     //command fail yes/no
-uint8_t canlisten = 0;
-uint8_t Lbuffer[8];         //8 byte buffer to store incomng can data
-uint8_t Gdata[9];
-uint8_t Glow, Ghigh;
+// uint8_t currentcanCommand;
+// uint8_t currentCanPage = 1;//Not the same as the speeduino config page numbers
+// uint8_t nCanretry = 0;      //no of retrys
+// uint8_t cancmdfail = 0;     //command fail yes/no
+// uint8_t canlisten = 0;
+// uint8_t Lbuffer[8];         //8 byte buffer to store incomng can data
+// uint8_t Gdata[9];
+// uint8_t Glow, Ghigh;
 
 #if defined(__AVR_ATmega1280__) || defined(__AVR_ATmega2560__)
-  HardwareSerial &CANSerial = Serial3;
+  extern HardwareSerial &CANSerial;
 #elif defined(CORE_STM32)
   #ifndef Serial2
     #define Serial2 Serial1
   #endif
   #if defined(STM32GENERIC) // STM32GENERIC core
-    SerialUART &CANSerial = Serial2;
+    extern SerialUART &CANSerial;
   #else //libmaple core aka STM32DUINO
-    HardwareSerial &CANSerial = Serial2;
+    extern HardwareSerial &CANSerial;
   #endif
 #elif defined(CORE_TEENSY)
-  HardwareSerial &CANSerial = Serial2;
+  extern HardwareSerial &CANSerial;
 #endif
 
 void canCommand();//This is the heart of the Command Line Interpeter.  All that needed to be done was to make it human readable.
