+++ speeduino/idle.h
@@ -28,39 +28,35 @@ struct StepperIdle
   byte moreAirDirection;
 };
 
-struct table2D iacClosedLoopTable;
-struct table2D iacPWMTable;
-struct table2D iacStepTable;
-//Open loop tables specifically for cranking
-struct table2D iacCrankStepsTable;
-struct table2D iacCrankDutyTable;
+// struct table2D iacClosedLoopTable;
+// struct table2D iacPWMTable;
+// struct table2D iacStepTable;
+// //Open loop tables specifically for cranking
+// struct table2D iacCrankStepsTable;
+// struct table2D iacCrankDutyTable;
 
-struct StepperIdle idleStepper;
-bool idleOn; //Simply tracks whether idle was on last time around
-byte idleInitComplete = 99; //TRacks which idle method was initialised. 99 is a method that will never exist
-unsigned int iacStepTime_uS;
-unsigned int iacCoolTime_uS;
-unsigned int completedHomeSteps;
+// struct StepperIdle idleStepper;
+// bool idleOn; //Simply tracks whether idle was on last time around
+// byte idleInitComplete = 99; //TRacks which idle method was initialised. 99 is a method that will never exist
+// unsigned int iacStepTime_uS;
+// unsigned int iacCoolTime_uS;
+// unsigned int completedHomeSteps;
 
-volatile PORT_TYPE *idle_pin_port;
-volatile PINMASK_TYPE idle_pin_mask;
-volatile PORT_TYPE *idle2_pin_port;
-volatile PINMASK_TYPE idle2_pin_mask;
+// volatile PORT_TYPE *idle_pin_port;
+// volatile PINMASK_TYPE idle_pin_mask;
+// volatile PORT_TYPE *idle2_pin_port;
+// volatile PINMASK_TYPE idle2_pin_mask;
 
-volatile bool idle_pwm_state;
-unsigned int idle_pwm_max_count; //Used for variable PWM frequency
-volatile unsigned int idle_pwm_cur_value;
-long idle_pid_target_value;
-long idle_pwm_target_value;
-long idle_cl_target_rpm;
-byte idleCounter; //Used for tracking the number of calls to the idle control function
+// volatile bool idle_pwm_state;
+// unsigned int idle_pwm_max_count; //Used for variable PWM frequency
+// volatile unsigned int idle_pwm_cur_value;
+// long idle_pid_target_value;
+// long idle_pwm_target_value;
+// long idle_cl_target_rpm;
+// byte idleCounter; //Used for tracking the number of calls to the idle control function
 
 void initialiseIdle();
-static inline void disableIdle();
-static inline void enableIdle();
-static inline byte isStepperHomed();
-static inline byte checkForStepping();
-static inline void doStep();
-static inline void idleInterrupt();
+void disableIdle();
+void idleInterrupt();
 
 #endif
