+++ speeduino/timers.h
@@ -19,22 +19,24 @@ Hence we will preload the timer with 131 cycles to leave 125 until overflow (1ms
 #ifndef TIMERS_H
 #define TIMERS_H
 
-volatile bool tachoAlt = false;
+#include <Arduino.h>
+
+// volatile bool tachoAlt = false;
 #define TACHO_PULSE_HIGH() *tach_pin_port |= (tach_pin_mask)
 #define TACHO_PULSE_LOW() *tach_pin_port &= ~(tach_pin_mask)
 enum TachoOutputStatus {DEACTIVE, READY, ACTIVE}; //The 3 statuses that the tacho output pulse can have
-volatile uint8_t tachoEndTime; //The time (in ms) that the tacho pulse needs to end at
-volatile TachoOutputStatus tachoOutputFlag;
+// volatile uint8_t tachoEndTime; //The time (in ms) that the tacho pulse needs to end at
+extern volatile TachoOutputStatus tachoOutputFlag;
 
-volatile byte loop33ms;
-volatile byte loop66ms;
-volatile byte loop100ms;
-volatile byte loop250ms;
-volatile int loopSec;
+// volatile byte loop33ms;
+// volatile byte loop66ms;
+// volatile byte loop100ms;
+// volatile byte loop250ms;
+// volatile int loopSec;
 
-volatile unsigned int dwellLimit_uS;
-volatile uint16_t lastRPM_100ms; //Need to record this for rpmDOT calculation
-volatile uint16_t last250msLoopCount = 1000; //Set to effectively random number on startup. Just need this to be different to what mainLoopCount equals initially (Probably 0)
+extern volatile unsigned int dwellLimit_uS;
+// volatile uint16_t lastRPM_100ms; //Need to record this for rpmDOT calculation
+// volatile uint16_t last250msLoopCount = 1000; //Set to effectively random number on startup. Just need this to be different to what mainLoopCount equals initially (Probably 0)
 
 #if defined (CORE_TEENSY)
   IntervalTimer lowResTimer;
